# Fix deep recursion on star end