--- a/ellipses/list_test.go
+++ b/ellipses/list_test.go
@@ -173,7 +173,7 @@ func TestFindListPatterns(t *testing.T) {
 					return s
 				}
 				if !reflect.DeepEqual(got, testCase.want) {
-					t.Errorf(fmt.Sprintf("want %s,", repl(testCase.want)))
+					t.Errorf("want %s,", repl(testCase.want))
 					t.Errorf("got %s,", repl(got))
 				}
 			}